requirements.txt
Pillow==10.0.0
opencv-python==4.8.0.76
numpy==1.24.3
matplotlib==3.7.1
librosa==0.10.1
scipy==1.10.1
Flask==2.0.1 